requirements.txt
youtube-transcript-api>=0.6.2
groq>=0.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
python-docx>=1.1.0
reportlab>=4.0.0
gradio>=4.0.0
